fix/moveFinderServiceToServiceDirectory: Move FinderHelper class from Helpers to Services The FinderService class has been relocated from the Helpers directory to the Services directory to better reflect its purpose and improve the project structure.
